+++ sql/core/src/test/scala/org/apache/spark/sql/execution/command/CharVarcharDDLTestBase.scala
@@ -17,10 +17,8 @@
 
 package org.apache.spark.sql.execution.command
 
-import org.apache.spark.{SparkConf, SparkException}
-import org.apache.spark.sql.{AnalysisException, DataFrame, QueryTest, Row}
-import org.apache.spark.sql.catalyst.expressions.RaiseError
-import org.apache.spark.sql.catalyst.plans.logical.Project
+import org.apache.spark.SparkConf
+import org.apache.spark.sql.{AnalysisException, QueryTest}
 import org.apache.spark.sql.catalyst.util.CharVarcharUtils
 import org.apache.spark.sql.connector.InMemoryPartitionTableCatalog
 import org.apache.spark.sql.internal.SQLConf
@@ -97,86 +95,6 @@ trait CharVarcharDDLTestBase extends QueryTest with SQLTestUtils {
       val v1 = e.getMessage contains "'VarcharType(4)' to 'c' with type 'VarcharType(3)'"
       val v2 = e.getMessage contains "varchar(4) cannot be cast to varchar(3)"
       assert(v1 || v2)
-    }
-  }
-
-  def checkRaiseErrorApplied(df: DataFrame): Unit = {
-    val re = df.queryExecution.optimizedPlan.asInstanceOf[Project].projectList.head.collect {
-      case _: RaiseError => true
-    }
-    assert(re.nonEmpty, "should apply raise_error to output")
-  }
-
-  test("create table w/ location and fit length values") {
-    Seq("char", "varchar").foreach { typ =>
-      withTempPath { dir =>
-        withTable("t") {
-          sql("SELECT '12' as c0").write.option("path", dir.toString).save()
-          sql(s"CREATE TABLE t (c0 $typ(2)) using $format LOCATION '$dir'")
-          val df = sql("select * from t")
-          checkRaiseErrorApplied(df)
-          // for non-v2 table we can track the result too
-          if (format != "foo") checkAnswer(df, Row("12"))
-        }
-      }
-    }
-  }
-
-  test("create table w/ location and over length values") {
-    Seq("char", "varchar").foreach { typ =>
-      withTempPath { dir =>
-        withTable("t") {
-          sql("SELECT '123456' as c0").write.option("path", dir.toString).save()
-          sql(s"CREATE TABLE t (c0 $typ(2)) using $format LOCATION '$dir'")
-          val df = sql("select * from t")
-          checkRaiseErrorApplied(df)
-
-          // for non-v2 table we can track the Runtime error too
-          if (format != "foo") {
-            val e = intercept[SparkException] { df.collect() }
-            assert(e.getCause.getMessage.contains(
-              s"input string of length 6 exceeds $typ type length limitation: 2"))
-          }
-        }
-      }
-    }
-  }
-
-  test("alter table set location w/ fit length values") {
-    Seq("char", "varchar").foreach { typ =>
-      withTempPath { dir =>
-        withTable("t") {
-          sql("SELECT '12' as c0").write.option("path", dir.toString).save()
-          sql(s"CREATE TABLE t (c0 $typ(2)) using $format")
-          sql(s"ALTER TABLE t SET LOCATION '$dir'")
-          val df = spark.table("t")
-          checkRaiseErrorApplied(df)
-
-          // for non-v2 table we can track the result too
-          if (format != "foo") checkAnswer(df, Row("12"))
-        }
-      }
-    }
-  }
-
-  test("alter table set location w/ over length values") {
-    Seq("char", "varchar").foreach { typ =>
-      withTempPath { dir =>
-        withTable("t") {
-          sql("SELECT '123456' as c0").write.option("path", dir.toString).save()
-          sql(s"CREATE TABLE t (c0 $typ(2)) using $format")
-          sql(s"ALTER TABLE t SET LOCATION '$dir'")
-          val df = spark.table("t")
-          checkRaiseErrorApplied(df)
-
-          // for non-v2 table we can track the Runtime error too
-          if (format != "foo") {
-            val e = intercept[SparkException] { df.collect() }
-            assert(e.getCause.getMessage.contains(
-              s"input string of length 6 exceeds $typ type length limitation: 2"))
-          }
-        }
-      }
     }
   }
 }
